@@ -1,4 +1,4 @@
-__version__ = "0.4.5.dev0"
+__version__ = "0.6.0a3"
 
 from . import buses
 from . import components
@@ -9,20 +9,26 @@ from . import processing
 from . import views
 from ._energy_system import EnergySystem
 from ._groupings import GROUPINGS
+from ._helpers import create_time_index
 from ._models import Model
 from ._options import Investment
 from ._options import NonConvex
 from ._plumbing import sequence
+from .buses import Bus  # default Bus (for convenience)
+from .flows import Flow  # default Flow (for convenience)
 
 __all__ = [
     "buses",
+    "Bus",
     "components",
     "constraints",
     "flows",
+    "Flow",
     "helpers",
     "processing",
     "views",
     "EnergySystem",
+    "create_time_index",
     "GROUPINGS",
     "Model",
     "Investment",
