@@ -5,27 +5,31 @@
 If you want to use solph on an energy system, you need to create it with these
 groupings specified like this:
 
-    .. code-block: python
+    .. code-block:: python
 
-    from oemof.network import EnergySystem
-    import solph
+        from oemof.network import EnergySystem
+        import oemof.solph as solph
 
-    energy_system = EnergySystem(groupings=solph.GROUPINGS)
+        energy_system = EnergySystem(groupings=solph.GROUPINGS)
 
 SPDX-FileCopyrightText: Uwe Krien <krien@uni-bremen.de>
 SPDX-FileCopyrightText: Simon Hilpert
 SPDX-FileCopyrightText: Cord Kaldemeyer
 SPDX-FileCopyrightText: Stephan Günther
+SPDX-FileCopyrightText: Saeed Sayadi
+SPDX-FileCopyrightText: Johannes Kochems
 
 SPDX-License-Identifier: MIT
-
 """
 
 from oemof.network import groupings as groupings
 
-from oemof.solph.flows._flow import FlowBlock
-from oemof.solph.flows._investment_flow import InvestmentFlowBlock
-from oemof.solph.flows._non_convex_flow import NonConvexFlowBlock
+from oemof.solph.flows._invest_non_convex_flow_block import (
+    InvestNonConvexFlowBlock,
+)
+from oemof.solph.flows._investment_flow_block import InvestmentFlowBlock
+from oemof.solph.flows._non_convex_flow_block import NonConvexFlowBlock
+from oemof.solph.flows._simple_flow_block import SimpleFlowBlock
 
 
 def constraint_grouping(node, fallback=lambda *xs, **ks: None):
@@ -58,15 +62,14 @@ def constraint_grouping(node, fallback=lambda *xs, **ks: None):
     return cg()
 
 
-standard_flow_grouping = groupings.FlowsWithNodes(constant_key=FlowBlock)
+standard_flow_grouping = groupings.FlowsWithNodes(constant_key=SimpleFlowBlock)
 
 
 def _investment_grouping(stf):
     if hasattr(stf[2], "investment"):
-        if stf[2].investment is not None:
+        if stf[2].investment is not None and stf[2].nonconvex is None:
             return True
-    else:
-        return False
+    return False
 
 
 investment_flow_grouping = groupings.FlowsWithNodes(
@@ -78,10 +81,9 @@ investment_flow_grouping = groupings.FlowsWithNodes(
 
 def _nonconvex_grouping(stf):
     if hasattr(stf[2], "nonconvex"):
-        if stf[2].nonconvex is not None:
+        if stf[2].nonconvex is not None and stf[2].investment is None:
             return True
-    else:
-        return False
+    return False
 
 
 nonconvex_flow_grouping = groupings.FlowsWithNodes(
@@ -89,9 +91,21 @@ nonconvex_flow_grouping = groupings.FlowsWithNodes(
 )
 
 
+def _invest_non_convex_grouping(stf):
+    if hasattr(stf[2], "nonconvex"):
+        if stf[2].investment is not None and stf[2].nonconvex is not None:
+            return True
+    return False
+
+
+invest_non_convex_flow_grouping = groupings.FlowsWithNodes(
+    constant_key=InvestNonConvexFlowBlock, filter=_invest_non_convex_grouping
+)
+
 GROUPINGS = [
     constraint_grouping,
     investment_flow_grouping,
     standard_flow_grouping,
     nonconvex_flow_grouping,
+    invest_non_convex_flow_grouping,
 ]
