@@ -11,6 +11,7 @@ SPDX-FileCopyrightText: Johannes Röder
 SPDX-FileCopyrightText: jakob-wo
 SPDX-FileCopyrightText: gplssm
 SPDX-FileCopyrightText: jnnr
+SPDX-FileCopyrightText: Johannes Kochems
 
 SPDX-License-Identifier: MIT
 
@@ -20,32 +21,49 @@ from oemof.solph.buses._bus import Bus
 
 
 class ElectricalBus(Bus):
-    r"""A electrical bus object. Every node has to be connected to BusBlock.
+    r"""An electrical bus object used for linear optimal power flow (LOPF)
+
+    Every (spatial) node has to be connected to a BusBlock.
     This BusBlock is used in combination with ElectricalLine objects
     for linear optimal power flow (lopf) calculations.
 
     Parameters
     ----------
     slack: boolean
-        If True BusBlock is slack bus for network
+        If True BusBlock is slack bus for electrical network
     v_max: numeric
         Maximum value of voltage angle at electrical bus
     v_min: numeric
-        Mininum value of voltag angle at electrical bus
+        Mininum value of voltage angle at electrical bus
 
     Note: This component is experimental. Use it with care.
 
     Notes
     -----
     The following sets, variables, constraints and objective parts are created
-     * :py:class:`~oemof.solph.blocks.bus.BusBlock`
+     * :py:class:`~oemof.solph._bus.BusBlock`
     The objects are also used inside:
      * :py:class:`~oemof.solph.experimental._electrical_line.ElectricalLine`
 
     """
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.slack = kwargs.get("slack", False)
-        self.v_max = kwargs.get("v_max", 1000)
-        self.v_min = kwargs.get("v_min", -1000)
+    def __init__(
+        self,
+        label=None,
+        *,
+        v_max,
+        v_min,
+        inputs=None,
+        outputs=None,
+        custom_properties=None,
+        slack=False,
+    ):
+        super().__init__(
+            label,
+            inputs=inputs,
+            outputs=outputs,
+            custom_properties=custom_properties,
+        )
+        self.slack = slack
+        self.v_max = v_max
+        self.v_min = v_min
