@@ -10,21 +10,31 @@ experimental code should be included in oemof.experimental.
 """
 
 from . import experimental
+from ._converter import Converter
+from ._converter import Transformer
 from ._extraction_turbine_chp import ExtractionTurbineCHP
 from ._generic_chp import GenericCHP
 from ._generic_storage import GenericStorage
-from ._offset_transformer import OffsetTransformer
+from ._link import Link
+from ._offset_converter import OffsetConverter
+from ._offset_converter import OffsetTransformer
+from ._offset_converter import slope_offset_from_nonconvex_input
+from ._offset_converter import slope_offset_from_nonconvex_output
 from ._sink import Sink
 from ._source import Source
-from ._transformer import Transformer
 
 __all__ = [
+    "Converter",
     "experimental",
     "ExtractionTurbineCHP",
     "GenericCHP",
     "GenericStorage",
+    "OffsetConverter",
+    "Link",
     "OffsetTransformer",
     "Sink",
-    "Source",
     "Transformer",
+    "Source",
+    "slope_offset_from_nonconvex_input",
+    "slope_offset_from_nonconvex_output",
 ]
