@@ -8,22 +8,48 @@ SPDX-FileCopyrightText: Simon Hilpert
 SPDX-FileCopyrightText: Cord Kaldemeyer
 SPDX-FileCopyrightText: Stephan Günther
 SPDX-FileCopyrightText: Birgit Schachler
+SPDX-FileCopyrightText: Johannes Kochems
 
 SPDX-License-Identifier: MIT
 
 """
-
-from oemof.network import network as on
-
-from oemof.solph._helpers import check_node_object_for_missing_attribute
+from oemof.network import Node
 
 
-class Sink(on.Sink):
-    """An object with one input flow."""
+class Sink(Node):
+    """A component which is designed for one input flow.
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        check_node_object_for_missing_attribute(self, "inputs")
+    Parameters
+    ----------
+    label : str or tuple
+        String holding the label of the Sink object.
+        The label of each object must be unique.
+    inputs: dict
+        A dictionary mapping input nodes to corresponding inflows
+        (i.e. input values).
+
+    Examples
+    --------
+    Defining a Sink:
+
+    >>> from oemof import solph
+    >>> bel = solph.buses.Bus(label='electricity')
+
+    >>> electricity_export = solph.components.Sink(
+    ...    label='el_export',
+    ...    inputs={bel: solph.flows.Flow()})
+
+    """
+
+    def __init__(self, label=None, *, inputs, custom_attributes=None):
+        if inputs is None:
+            inputs = {}
+        if custom_attributes is None:
+            custom_attributes = {}
+
+        super().__init__(
+            label=label, inputs=inputs, custom_properties=custom_attributes
+        )
 
     def constraint_group(self):
         pass
