@@ -38,14 +38,16 @@ def shared_limit(
         Name of the constraint to create
     quantity : pyomo.core.base.var.IndexedVar
         Shared Pyomo variable for all components of a type.
+        (:math:`v_i(t)`)
     components : list of components
         list of components of the same type
     weights : list of numeric values
         has to have the same length as the list of components
+        (:math:`w_i(t)`)
     lower_limit : numeric
-        the lower limit
+        the lower limit (:math:`l_\mathrm{low}`)
     upper_limit : numeric
-        the lower limit
+        the upper limit (:math:`l_\mathrm{up}`)
 
     Examples
     --------
@@ -58,20 +60,25 @@ def shared_limit(
 
     >>> import pandas as pd
     >>> from oemof import solph
-    >>> date_time_index = pd.date_range('1/1/2012', periods=5, freq='H')
-    >>> energysystem = solph.EnergySystem(timeindex=date_time_index)
+    >>> date_time_index = pd.date_range('1/1/2012', periods=6, freq='h')
+    >>> energysystem = solph.EnergySystem(
+    ...     timeindex=date_time_index,
+    ...     infer_last_interval=False,
+    ... )
     >>> b1 = solph.buses.Bus(label="Party1Bus")
     >>> b2 = solph.buses.Bus(label="Party2Bus")
     >>> storage1 = solph.components.GenericStorage(
     ...     label="Party1Storage",
-    ...     nominal_storage_capacity=5,
+    ...     nominal_capacity=5,
     ...     inputs={b1: solph.flows.Flow()},
-    ...     outputs={b1: solph.flows.Flow()})
+    ...     outputs={b1: solph.flows.Flow()}
+    ... )
     >>> storage2 = solph.components.GenericStorage(
     ...     label="Party2Storage",
-    ...     nominal_storage_capacity=5,
+    ...     nominal_capacity=5,
     ...     inputs={b1: solph.flows.Flow()},
-    ...     outputs={b1: solph.flows.Flow()})
+    ...     outputs={b1: solph.flows.Flow()}
+    ... )
     >>> energysystem.add(b1, b2, storage1, storage2)
     >>> components = [storage1, storage2]
     >>> model = solph.Model(energysystem)
@@ -79,7 +86,8 @@ def shared_limit(
     ...    model,
     ...    model.GenericStorageBlock.storage_content,
     ...    "limit_storage", components,
-    ...    [1, 1], upper_limit=5)
+    ...    [1, 1], upper_limit=5
+    ... )
     """
 
     setattr(model, limit_name, po.Var(model.TIMESTEPS))
