@@ -2,8 +2,7 @@
 
 """
 This is a collection of helper functions which work on their own and can be
-used by various classes. If there are too many helper-functions, they will
-be sorted in different modules.
+used by various classes.
 
 SPDX-FileCopyrightText: Uwe Krien <krien@uni-bremen.de>
 SPDX-FileCopyrightText: Caroline Möller
@@ -16,13 +15,8 @@ SPDX-License-Identifier: MIT
 
 """
 
-import datetime as dt
 import os
 from collections.abc import MutableMapping
-
-import pandas as pd
-
-from oemof.solph._plumbing import sequence
 
 
 def get_basic_path():
@@ -52,7 +46,8 @@ def flatten(d, parent_key="", sep="_"):
     See: https://stackoverflow.com/questions/6027558/
          flatten-nested-python-dictionaries-compressing-keys
 
-    d : dictionary
+    d : (dictionary)
+    parent_key : (do not use this, used internally for recursion)
     sep : separator for flattening keys
 
     Returns
@@ -67,38 +62,3 @@ def flatten(d, parent_key="", sep="_"):
         else:
             items.append((new_key, v))
     return dict(items)
-
-
-def calculate_timeincrement(timeindex, fill_value=None):
-    """
-    Calculates timeincrement for `timeindex`
-
-    Parameters
-    ----------
-    timeindex: pd.DatetimeIndex
-        timeindex of energysystem
-    fill_value: numerical
-        timeincrement for first timestep in hours
-    """
-    if isinstance(timeindex, pd.DatetimeIndex) and (
-        fill_value
-        and isinstance(fill_value, pd.Timedelta)
-        or fill_value is None
-    ):
-        if len(set(timeindex)) != len(timeindex):
-            raise IndexError("No equal DatetimeIndex allowed!")
-        timeindex = timeindex.to_series()
-        timeindex_sorted = timeindex.sort_values()
-        if fill_value:
-            timeincrement = timeindex_sorted.diff().fillna(value=fill_value)
-        else:
-            timeincrement = timeindex_sorted.diff().fillna(method="bfill")
-        timeincrement_sec = timeincrement.map(dt.timedelta.total_seconds)
-        timeincrement_hourly = list(timeincrement_sec.map(lambda x: x / 3600))
-        timeincrement = sequence(timeincrement_hourly)
-        return timeincrement
-    else:
-        raise AttributeError(
-            "'timeindex' must be of type 'DatetimeIndex' and "
-            + "'fill_value' of type 'Timedelta'."
-        )
