@@ -9,25 +9,29 @@ SPDX-FileCopyrightText: Simon Hilpert
 SPDX-FileCopyrightText: Cord Kaldemeyer
 SPDX-FileCopyrightText: Stephan Günther
 SPDX-FileCopyrightText: henhuy
+SPDX-FileCopyrightText: Johannes Kochems
+SPDX-FileCopyrightText: Patrik Schönfeldt <patrik.schoenfeldt@dlr.de>
 
 SPDX-License-Identifier: MIT
 
 """
 
 import sys
+from collections import abc
 from itertools import groupby
 
+import numpy as np
 import pandas as pd
-from oemof.network.network import Node
+from oemof.network.network import Entity
 from pyomo.core.base.piecewise import IndexedPiecewise
 from pyomo.core.base.var import Var
 
+from ._plumbing import _FakeSequence
 from .helpers import flatten
 
 
 def get_tuple(x):
-    """
-    Get oemof tuple within iterable or create it.
+    """Get oemof tuple within iterable or create it
 
     Tuples from Pyomo are of type `(n, n, int)`, `(n, n)` and `(n, int)`.
     For single nodes `n` a tuple with one object `(n,)` is created.
@@ -35,7 +39,7 @@ def get_tuple(x):
     for i in x:
         if isinstance(i, tuple):
             return i
-        elif issubclass(type(i), Node):
+        elif issubclass(type(i), Entity):
             return (i,)
 
     # for standalone variables, x is used as identifying tuple
@@ -44,38 +48,35 @@ def get_tuple(x):
 
 
 def get_timestep(x):
-    """
-    Get the timestep from oemof tuples.
+    """Get the timestep from oemof tuples
 
     The timestep from tuples `(n, n, int)`, `(n, n)`, `(n, int)` and (n,)
     is fetched as the last element. For time-independent data (scalars)
     zero ist returned.
     """
-    if all(issubclass(type(n), Node) for n in x):
+    if all(issubclass(type(n), Entity) for n in x):
         return 0
     else:
         return x[-1]
 
 
 def remove_timestep(x):
-    """
-    Remove the timestep from oemof tuples.
+    """Remove the timestep from oemof tuples
 
     The timestep is removed from tuples of type `(n, n, int)` and `(n, int)`.
     """
-    if all(issubclass(type(n), Node) for n in x):
+    if all(issubclass(type(n), Entity) for n in x):
         return x
     else:
         return x[:-1]
 
 
 def create_dataframe(om):
-    """
-    Create a result dataframe with all optimization data.
+    """Create a result DataFrame with all optimization data
 
-    Results from Pyomo are written into pandas DataFrame where separate columns
-    are created for the variable index e.g. for tuples of the flows and
-    components or the timesteps.
+    Results from Pyomo are written into one common pandas.DataFrame where
+    separate columns are created for the variable index e.g. for tuples
+    of the flows and components or the timesteps.
     """
     # get all pyomo variables including their block
     block_vars = list(
@@ -86,7 +87,13 @@ def create_dataframe(om):
         # Drop the auxiliary variables introduced by pyomo's Piecewise
         parent_component = bv.parent_block().parent_component()
         if not isinstance(parent_component, IndexedPiecewise):
-            for i in getattr(bv, "_index"):
+            try:
+                idx_set = getattr(bv, "_index_set")
+            except AttributeError:
+                # To make it compatible with Pyomo < 6.4.1
+                idx_set = getattr(bv, "_index")
+
+            for i in idx_set:
                 key = (str(bv).split(".")[0], str(bv).split(".")[-1], i)
                 value = bv[i].value
                 var_dict[key] = value
@@ -103,6 +110,11 @@ def create_dataframe(om):
     df["timestep"] = df["oemof_tuple"].map(get_timestep)
     df["oemof_tuple"] = df["oemof_tuple"].map(remove_timestep)
 
+    # Use another call of remove timestep to get rid of period not needed
+    df.loc[df["variable_name"] == "flow", "oemof_tuple"] = df.loc[
+        df["variable_name"] == "flow", "oemof_tuple"
+    ].map(remove_timestep)
+
     # order the data by oemof tuple and timestep
     df = df.sort_values(["oemof_tuple", "timestep"], ascending=[True, True])
 
@@ -112,72 +124,299 @@ def create_dataframe(om):
     return df
 
 
-def results(om):
+def divide_scalars_sequences(df_dict, k):
+    """Split results into scalars and sequences results
+
+    Parameters
+    ----------
+    df_dict: dict
+        dict of pd.DataFrames, keyed by oemof tuples
+    k: tuple
+        oemof tuple for results processing
     """
-    Create a result dictionary from the result DataFrame.
+    try:
+        condition = df_dict[k][:-1].isnull().any()
+        scalars = df_dict[k].loc[:, condition].dropna().iloc[0]
+        sequences = df_dict[k].loc[:, ~condition]
+        return {"scalars": scalars, "sequences": sequences}
+    except IndexError:
+        error_message = (
+            "Cannot access index on result data. "
+            + "Did the optimization terminate"
+            + " without errors?"
+        )
+        raise IndexError(error_message)
 
-    Results from Pyomo are written into a dictionary of pandas objects where
-    a Series holds all scalar values and a dataframe all sequences for nodes
-    and flows.
-    The dictionary is keyed by the nodes e.g. `results[idx]['scalars']`
-    and flows e.g. `results[n, n]['sequences']`.
+
+def set_result_index(df_dict, k, result_index):
+    """Define index for results
+
+    Parameters
+    ----------
+    df_dict: dict
+        dict of pd.DataFrames, keyed by oemof tuples
+    k: tuple
+        oemof tuple for results processing
+    result_index: pd.Index
+        Index to use for results
     """
-    df = create_dataframe(om)
-
-    # create a dict of dataframes keyed by oemof tuples
-    df_dict = {
-        k
-        if len(k) > 1
-        else (k[0], None): v[["timestep", "variable_name", "value"]]
-        for k, v in df.groupby("oemof_tuple")
-    }
-
-    # create final result dictionary by splitting up the dataframes in the
-    # dataframe dict into a series for scalar data and dataframe for sequences
-    result = {}
-    for k in df_dict:
-        df_dict[k].set_index("timestep", inplace=True)
-        df_dict[k] = df_dict[k].pivot(columns="variable_name", values="value")
+    try:
+        df_dict[k].index = result_index
+    except ValueError:
         try:
-            df_dict[k].index = om.es.timeindex
+            df_dict[k] = df_dict[k][:-1]
+            df_dict[k].index = result_index
         except ValueError as e:
             msg = (
-                "\nFlowBlock: {0}-{1}. This could be caused by NaN-values in"
-                " your input data."
+                "\nFlow: {0}-{1}. This could be caused by NaN-values "
+                "in your input data."
             )
             raise type(e)(
                 str(e) + msg.format(k[0].label, k[1].label)
             ).with_traceback(sys.exc_info()[2])
+
+
+def set_sequences_index(df, result_index):
+    try:
+        df.index = result_index
+    except ValueError:
         try:
-            condition = df_dict[k].isnull().any()
-            scalars = df_dict[k].loc[:, condition].dropna().iloc[0]
-            sequences = df_dict[k].loc[:, ~condition]
-            result[k] = {"scalars": scalars, "sequences": sequences}
-        except IndexError:
-            error_message = (
-                "Cannot access index on result data. "
-                + "Did the optimization terminate"
-                + " without errors?"
-            )
-            raise IndexError(error_message)
+            df = df[:-1]
+            df.index = result_index
+        except ValueError:
+            raise ValueError("Results extraction failed!")
+
+
+def results(model, remove_last_time_point=False):
+    """Create a nested result dictionary from the result DataFrame
+
+    The already rearranged results from Pyomo from the result DataFrame are
+    transferred into a nested dictionary of pandas objects.
+    The first level key of that dictionary is a node (denoting the respective
+    flow or component).
+
+    The second level keys are "sequences" and "scalars" for a *standard model*:
+
+    * A pd.DataFrame holds all results that are time-dependent, i.e. given as
+      a sequence and can be indexed with the energy system's timeindex.
+    * A pd.Series holds all scalar values which are applicable for timestep 0
+      (i.e. investments).
+
+    For a *multi-period model*, the second level key for "sequences" remains
+    the same while instead of "scalars", the key "period_scalars" is used:
+
+    * For sequences, see standard model.
+    * Instead of a pd.Series, a pd.DataFrame holds scalar values indexed
+      by periods. These hold investment-related variables.
+
+    Examples
+    --------
+    * *Standard model*: `results[idx]['scalars']`
+      and flows `results[n, n]['sequences']`.
+    * *Multi-period model*: `results[idx]['period_scalars']`
+      and flows `results[n, n]['sequences']`.
+
+    Parameters
+    ----------
+    model : oemof.solph.Model
+        A solved oemof.solph model.
+    remove_last_time_point : bool
+        The last time point of all TIMEPOINT variables is removed to get the
+        same length as the TIMESTEP (interval) variables without getting
+        nan-values. By default, the last time point is removed if it has not
+        been defined by the user in the EnergySystem but inferred. If all
+        time points have been defined explicitly by the user the last time
+        point will not be removed by default. In that case all interval
+        variables will get one row with nan-values to have the same index
+        for all variables.
+    """
+    # Extraction steps that are the same for both model types
+    df = create_dataframe(model)
+
+    # create a dict of dataframes keyed by oemof tuples
+    df_dict = {
+        k if len(k) > 1 else (k[0], None): v[
+            ["timestep", "variable_name", "value"]
+        ]
+        for k, v in df.groupby("oemof_tuple")
+    }
+
+    # Define index
+    if model.es.timeindex is None:
+        result_index = list(range(len(model.es.timeincrement) + 1))
+    else:
+        result_index = model.es.timeindex
+
+    # create final result dictionary by splitting up the dataframes in the
+    # dataframe dict into a series for scalar data and dataframe for sequences
+    result = {}
+
+    # Standard model results extraction
+    if model.es.periods is None:
+        result = _extract_standard_model_result(
+            df_dict, result, result_index, remove_last_time_point
+        )
+        scalars_col = "scalars"
+
+    # Results extraction for a multi-period model
+    else:
+        period_indexed = ["invest", "total", "old", "old_end", "old_exo"]
+
+        result = _extract_multi_period_model_result(
+            model,
+            df_dict,
+            period_indexed,
+            result,
+            result_index,
+            remove_last_time_point,
+        )
+        scalars_col = "period_scalars"
 
     # add dual variables for bus constraints
-    if om.dual is not None:
+    if model.dual is not None:
         grouped = groupby(
-            sorted(om.BusBlock.balance.iterkeys()), lambda p: p[0]
+            sorted(model.BusBlock.balance.iterkeys()), lambda t: t[0]
         )
-        for bus, timesteps in grouped:
+        for bus, timestep in grouped:
             duals = [
-                om.dual[om.BusBlock.balance[bus, t]] for _, t in timesteps
+                model.dual[model.BusBlock.balance[bus, t]] for _, t in timestep
             ]
-            df = pd.DataFrame({"duals": duals}, index=om.es.timeindex)
+            if model.es.periods is None:
+                df = pd.DataFrame({"duals": duals}, index=result_index[:-1])
+            # TODO: Align with standard model
+            else:
+                df = pd.DataFrame({"duals": duals}, index=result_index)
             if (bus, None) not in result.keys():
                 result[(bus, None)] = {
                     "sequences": df,
-                    "scalars": pd.Series(dtype=float),
+                    scalars_col: pd.Series(dtype=float),
                 }
             else:
                 result[(bus, None)]["sequences"]["duals"] = duals
+
+    return result
+
+
+def _extract_standard_model_result(
+    df_dict, result, result_index, remove_last_time_point
+):
+    """Extract and return the results of a standard model
+
+    * Optionally remove last time point or include it elsewise.
+    * Set index to timeindex and pivot results such that values are displayed
+      for the respective variables. Reindex with the energy system's timeindex.
+    * Filter for columns with nan values to retrieve scalar variables. Split
+      up the DataFrame into sequences and scalars and return it.
+
+    Parameters
+    ----------
+    df_dict : dict
+        dictionary of results DataFrames
+    result : dict
+        dictionary to store the results
+    result_index : pd.DatetimeIndex
+        timeindex to use for the results (derived from EnergySystem)
+    remove_last_time_point : bool
+        if True, remove the last time point
+
+    Returns
+    -------
+    result : dict
+        dictionary with results stored
+    """
+    if remove_last_time_point:
+        # The values of intervals belong to the time at the beginning of the
+        # interval.
+        for k in df_dict:
+            df_dict[k].set_index("timestep", inplace=True)
+            df_dict[k] = df_dict[k].pivot(
+                columns="variable_name", values="value"
+            )
+            set_result_index(df_dict, k, result_index[:-1])
+            result[k] = divide_scalars_sequences(df_dict, k)
+    else:
+        for k in df_dict:
+            df_dict[k].set_index("timestep", inplace=True)
+            df_dict[k] = df_dict[k].pivot(
+                columns="variable_name", values="value"
+            )
+            # Add empty row with nan at the end of the table by adding 1 to the
+            # last value of the numeric index.
+            df_dict[k].loc[df_dict[k].index[-1] + 1, :] = np.nan
+            set_result_index(df_dict, k, result_index)
+            result[k] = divide_scalars_sequences(df_dict, k)
+
+    return result
+
+
+def _extract_multi_period_model_result(
+    model,
+    df_dict,
+    period_indexed=None,
+    result=None,
+    result_index=None,
+    remove_last_time_point=False,
+):
+    """Extract and return the results of a multi-period model
+
+    Difference to standard model is in the way, scalar values are extracted
+    since they now depend on periods.
+
+    Parameters
+    ----------
+    model : oemof.solph.models.Model
+        The optimization model
+    df_dict : dict
+        dictionary of results DataFrames
+    period_indexed : list
+        list of variables that are indexed by periods
+    result : dict
+        dictionary to store the results
+    result_index : pd.DatetimeIndex
+        timeindex to use for the results (derived from EnergySystem)
+    remove_last_time_point : bool
+        if True, remove the last time point
+
+    Returns
+    -------
+    result : dict
+        dictionary with results stored
+    """
+    for k in df_dict:
+        df_dict[k].set_index("timestep", inplace=True)
+        df_dict[k] = df_dict[k].pivot(columns="variable_name", values="value")
+        # Split data set
+        period_cols = [
+            col for col in df_dict[k].columns if col in period_indexed
+        ]
+        # map periods to their start years for displaying period results
+        d = {
+            key: val + model.es.periods[0].min().year
+            for key, val in enumerate(model.es.periods_years)
+        }
+        period_scalars = df_dict[k].loc[:, period_cols].dropna()
+        sequences = df_dict[k].loc[
+            :, [col for col in df_dict[k].columns if col not in period_cols]
+        ]
+        if remove_last_time_point:
+            set_sequences_index(sequences, result_index[:-1])
+        else:
+            set_sequences_index(sequences, result_index)
+        if period_scalars.empty:
+            period_scalars = pd.DataFrame(index=d.values())
+        try:
+            period_scalars.rename(index=d, inplace=True)
+            period_scalars.index.name = "period"
+            result[k] = {
+                "period_scalars": period_scalars,
+                "sequences": sequences,
+            }
+        except IndexError:
+            error_message = (
+                "Some indices seem to be not matching.\n"
+                "Cannot properly extract model results."
+            )
+            raise IndexError(error_message)
 
     return result
 
@@ -192,11 +431,11 @@ def convert_keys_to_strings(result, keep_none_type=False):
     """
     if keep_none_type:
         converted = {
-            tuple([str(e) if e is not None else None for e in k])
-            if isinstance(k, tuple)
-            else str(k)
-            if k is not None
-            else None: v
+            (
+                tuple([str(e) if e is not None else None for e in k])
+                if isinstance(k, tuple)
+                else str(k) if k is not None else None
+            ): v
             for k, v in result.items()
         }
     else:
@@ -209,7 +448,7 @@ def convert_keys_to_strings(result, keep_none_type=False):
 
 def meta_results(om, undefined=False):
     """
-    Fetch some meta data from the Solver. Feel free to add more keys.
+    Fetch some metadata from the Solver. Feel free to add more keys.
 
     Valid keys of the resulting dictionary are: 'objective', 'problem',
     'solver'.
@@ -273,7 +512,8 @@ def __separate_attrs(
     """
 
     def detect_scalars_and_sequences(com):
-        com_data = {"scalars": {}, "sequences": {}}
+        scalars = {}
+        sequences = {}
 
         default_exclusions = [
             "__",
@@ -291,7 +531,7 @@ def __separate_attrs(
         attrs = [
             i
             for i in dir(com)
-            if not (callable(getattr(com, i)) or i.startswith(exclusions))
+            if not (i.startswith(exclusions) or callable(getattr(com, i)))
         ]
 
         for a in attrs:
@@ -301,13 +541,13 @@ def __separate_attrs(
             # "investment" prefix to component data:
             if attr_value.__class__.__name__ == "Investment":
                 invest_data = detect_scalars_and_sequences(attr_value)
-                com_data["scalars"].update(
+                scalars.update(
                     {
                         "investment_" + str(k): v
                         for k, v in invest_data["scalars"].items()
                     }
                 )
-                com_data["sequences"].update(
+                sequences.update(
                     {
                         "investment_" + str(k): v
                         for k, v in invest_data["sequences"].items()
@@ -316,24 +556,27 @@ def __separate_attrs(
                 continue
 
             if isinstance(attr_value, str):
-                com_data["scalars"][a] = attr_value
+                scalars[a] = attr_value
                 continue
 
             # If the label is a tuple it is iterable, therefore it should be
-            # converted to a string. Otherwise it will be a sequence.
+            # converted to a string. Otherwise, it will be a sequence.
             if a == "label":
                 attr_value = str(attr_value)
 
-            # check if attribute is iterable
-            # see: https://stackoverflow.com/questions/1952464/
-            # in-python-how-do-i-determine-if-an-object-is-iterable
-            try:
-                _ = (e for e in attr_value)
-                com_data["sequences"][a] = attr_value
-            except TypeError:
-                com_data["scalars"][a] = attr_value
+            if isinstance(attr_value, abc.Iterable):
+                sequences[a] = attr_value
+            elif isinstance(attr_value, _FakeSequence):
+                scalars[a] = attr_value.value
+            else:
+                scalars[a] = attr_value
 
-        com_data["sequences"] = flatten(com_data["sequences"])
+        sequences = flatten(sequences)
+
+        com_data = {
+            "scalars": scalars,
+            "sequences": sequences,
+        }
         move_undetected_scalars(com_data)
         if exclude_none:
             remove_nones(com_data)
@@ -349,19 +592,11 @@ def __separate_attrs(
             if isinstance(value, str):
                 com["scalars"][ckey] = value
                 del com["sequences"][ckey]
-                continue
-            try:
-                _ = (e for e in value)
-            except TypeError:
-                com["scalars"][ckey] = value
+            elif isinstance(value, _FakeSequence):
+                com["scalars"][ckey] = value.value
                 del com["sequences"][ckey]
-            else:
-                try:
-                    if not value.default_changed:
-                        com["scalars"][ckey] = value.default
-                        del com["sequences"][ckey]
-                except AttributeError:
-                    pass
+            elif len(value) == 0:
+                del com["sequences"][ckey]
 
     def remove_nones(com):
         for ckey, value in list(com["scalars"].items()):
