@@ -10,9 +10,7 @@ SPDX-License-Identifier: MIT
 """
 
 import pandas
-from nose.tools import assert_raises
-from nose.tools import eq_
-from nose.tools import ok_
+import pytest
 from pandas.testing import assert_frame_equal
 from pandas.testing import assert_series_equal
 
@@ -22,9 +20,9 @@ from oemof.solph import Model
 from oemof.solph import processing
 from oemof.solph import views
 from oemof.solph.buses import Bus
+from oemof.solph.components import Converter
 from oemof.solph.components import GenericStorage
 from oemof.solph.components import Sink
-from oemof.solph.components import Transformer
 from oemof.solph.flows import Flow
 
 
@@ -34,8 +32,11 @@ class TestParameterResult:
         cls.period = 24
         cls.es = EnergySystem(
             timeindex=pandas.date_range(
-                "2016-01-01", periods=cls.period, freq="H"
-            )
+                "2016-01-01",
+                periods=cls.period,
+                freq="h",
+            ),
+            infer_last_interval=True,
         )
 
         # BUSSES
@@ -45,12 +46,12 @@ class TestParameterResult:
         cls.es.add(b_el1, b_el2, b_diesel)
 
         # TEST DIESEL:
-        dg = Transformer(
+        dg = Converter(
             label="diesel",
             inputs={b_diesel: Flow(variable_costs=2)},
             outputs={
                 b_el1: Flow(
-                    variable_costs=1, investment=Investment(ep_costs=0.5)
+                    variable_costs=1, nominal_capacity=Investment(ep_costs=0.5)
                 )
             },
             conversion_factors={b_el1: 2},
@@ -66,7 +67,7 @@ class TestParameterResult:
             invest_relation_output_capacity=1 / 6,
             inflow_conversion_factor=1,
             outflow_conversion_factor=0.8,
-            investment=Investment(ep_costs=0.4),
+            nominal_capacity=Investment(ep_costs=0.4),
         )
 
         cls.demand_values = [0.0] + [100] * 23
@@ -74,7 +75,7 @@ class TestParameterResult:
             label="demand_el",
             inputs={
                 b_el2: Flow(
-                    nominal_value=1,
+                    nominal_capacity=1,
                     fix=cls.demand_values,
                 )
             },
@@ -96,8 +97,10 @@ class TestParameterResult:
             param_results[(b_el2, demand)]["scalars"].sort_index(),
             pandas.Series(
                 {
+                    "bidirectional": False,
                     "fixed": True,
-                    "nominal_value": 1,
+                    "integer": False,
+                    "nominal_capacity": 1,
                     "max": 1,
                     "min": 0,
                     "variable_costs": 0,
@@ -117,26 +120,30 @@ class TestParameterResult:
         param_results = processing.parameter_as_dict(
             self.es, exclude_none=False
         )
-        scalar_attributes = {
-            'fixed': True,
-            "integer": None,
+        default_attributes = {
+            "age": None,
+            "fixed": True,
+            "lifetime": None,
+            "integer": False,
             "investment": None,
-            "nominal_value": 1,
+            "nominal_capacity": 1,
             "nonconvex": None,
+            "bidirectional": False,
             "full_load_time_max": None,
             "full_load_time_min": None,
             "max": 1,
             "min": 0,
-            "negative_gradient_ub": None,
-            "positive_gradient_ub": None,
+            "negative_gradient_limit": None,
+            "positive_gradient_limit": None,
             "variable_costs": 0,
+            "fixed_costs": None,
             "flow": None,
             "values": None,
             "label": str(b_el2.outputs[demand].label),
         }
         assert_series_equal(
             param_results[(b_el2, demand)]["scalars"].sort_index(),
-            pandas.Series(scalar_attributes).sort_index(),
+            pandas.Series(default_attributes).sort_index(),
         )
         sequences_attributes = {
             "value": self.demand_values,
@@ -164,18 +171,22 @@ class TestParameterResult:
                 {
                     "balanced": True,
                     "initial_storage_level": 0,
-                    "invest_relation_input_capacity": 1 / 6,
-                    "invest_relation_output_capacity": 1 / 6,
-                    "investment_ep_costs": 0.4,
+                    "investment_age": 0,
                     "investment_existing": 0,
+                    "investment_interest_rate": 0,
+                    "investment_nonconvex": False,
+                    "investment_ep_costs": 0.4,
                     "investment_maximum": float("inf"),
                     "investment_minimum": 0,
                     "investment_nonconvex": False,
                     "investment_offset": 0,
                     "label": "storage",
+                    "fixed_costs": 0,
                     "fixed_losses_absolute": 0,
                     "fixed_losses_relative": 0,
                     "inflow_conversion_factor": 1,
+                    "invest_relation_input_capacity": 1 / 6,
+                    "invest_relation_output_capacity": 1 / 6,
                     "loss_rate": 0,
                     "max_storage_level": 1,
                     "min_storage_level": 0,
@@ -200,18 +211,22 @@ class TestParameterResult:
                 {
                     "balanced": True,
                     "initial_storage_level": 0,
-                    "invest_relation_input_capacity": 1 / 6,
-                    "invest_relation_output_capacity": 1 / 6,
-                    "investment_ep_costs": 0.4,
+                    "investment_age": 0,
                     "investment_existing": 0,
+                    "investment_interest_rate": 0,
+                    "investment_nonconvex": False,
+                    "investment_ep_costs": 0.4,
                     "investment_maximum": float("inf"),
                     "investment_minimum": 0,
                     "investment_nonconvex": False,
                     "investment_offset": 0,
                     "label": "storage",
+                    "fixed_costs": 0,
                     "fixed_losses_absolute": 0,
                     "fixed_losses_relative": 0,
                     "inflow_conversion_factor": 1,
+                    "invest_relation_input_capacity": 1 / 6,
+                    "invest_relation_output_capacity": 1 / 6,
                     "loss_rate": 0,
                     "max_storage_level": 1,
                     "min_storage_level": 0,
@@ -270,49 +285,52 @@ class TestParameterResult:
         )
 
         bel1_m = views.node(param_results, "b_el1", multiindex=True)
-        eq_(bel1_m["scalars"][("b_el1", "storage", "variable_costs")], 3)
+        assert bel1_m["scalars"][("b_el1", "storage", "variable_costs")] == 3
 
     def test_multiindex_sequences(self):
         results = processing.results(self.om)
         bel1 = views.node(results, "b_el1", multiindex=True)
-        eq_(int(bel1["sequences"][("diesel", "b_el1", "flow")].sum()), 2875)
+        assert (
+            int(bel1["sequences"][("diesel", "b_el1", "flow")].sum()) == 2875
+        )
 
     def test_error_from_nan_values(self):
         trsf = self.es.groups["diesel"]
         bus = self.es.groups["b_el1"]
         self.mod.flow[trsf, bus, 5] = float("nan")
-        with assert_raises(ValueError):
+        with pytest.raises(ValueError):
             processing.results(self.mod)
 
     def test_duals(self):
         results = processing.results(self.om)
         bel = views.node(results, "b_el1", multiindex=True)
-        eq_(int(bel["sequences"]["b_el1", "None", "duals"].sum()), 48)
+        assert int(bel["sequences"]["b_el1", "None", "duals"].sum()) == 48
 
     def test_node_weight_by_type(self):
         results = processing.results(self.om)
         storage_content = views.node_weight_by_type(
             results, node_type=GenericStorage
         )
-        eq_(round(float(storage_content.sum()), 6), 1437.500003)
+        assert (
+            storage_content.sum().iloc[0] == pytest.approx(1437.5, abs=0.1)
+        ).all()
 
     def test_output_by_type_view(self):
         results = processing.results(self.om)
-        transformer_output = views.node_output_by_type(
-            results, node_type=Transformer
+        converter_output = views.node_output_by_type(
+            results, node_type=Converter
         )
         compare = views.node(results, "diesel", multiindex=True)["sequences"][
             ("diesel", "b_el1", "flow")
         ]
-        eq_(int(transformer_output.sum()), int(compare.sum()))
+        assert converter_output.sum().iloc[0] == pytest.approx(compare.sum())
 
     def test_input_by_type_view(self):
         results = processing.results(self.om)
         sink_input = views.node_input_by_type(results, node_type=Sink)
         compare = views.node(results, "demand_el", multiindex=True)
-        eq_(
-            int(sink_input.sum()),
-            int(compare["sequences"][("b_el2", "demand_el", "flow")].sum()),
+        assert sink_input.sum().iloc[0] == pytest.approx(
+            compare["sequences"][("b_el2", "demand_el", "flow")].sum()
         )
 
     def test_net_storage_flow(self):
@@ -320,36 +338,41 @@ class TestParameterResult:
         storage_flow = views.net_storage_flow(
             results, node_type=GenericStorage
         )
+
         compare = views.node(results, "storage", multiindex=True)["sequences"]
-        eq_(
+
+        assert (
             (
                 (
                     compare[("storage", "b_el2", "flow")]
                     - compare[("b_el1", "storage", "flow")]
-                ).to_frame()
+                )
+                .to_frame()
+                .fillna(0)
                 == storage_flow.values
-            ).all()[0],
-            True,
+            )
+            .all()
+            .iloc[0]
         )
 
     def test_output_by_type_view_empty(self):
         results = processing.results(self.om)
         view = views.node_output_by_type(results, node_type=Flow)
-        ok_(view is None)
+        assert view is None
 
     def test_input_by_type_view_empty(self):
         results = processing.results(self.om)
         view = views.node_input_by_type(results, node_type=Flow)
-        ok_(view is None)
+        assert view is None
 
     def test_net_storage_flow_empty(self):
         results = processing.results(self.om)
         view = views.net_storage_flow(results, node_type=Sink)
-        ok_(view is None)
+        assert view is None
         view2 = views.net_storage_flow(results, node_type=Flow)
-        ok_(view2 is None)
+        assert view2 is None
 
     def test_node_weight_by_type_empty(self):
         results = processing.results(self.om)
         view = views.node_weight_by_type(results, node_type=Flow)
-        ok_(view is None)
+        assert view is None
