@@ -14,8 +14,7 @@ import logging
 import os
 
 import pandas as pd
-from nose.tools import eq_
-from oemof.network import network
+import pytest
 
 from oemof.solph import EnergySystem
 from oemof.solph import Investment
@@ -29,10 +28,9 @@ from oemof.solph.flows import Flow
 
 
 def test_connect_invest():
-    date_time_index = pd.date_range("1/1/2012", periods=24 * 7, freq="H")
+    date_time_index = pd.date_range("1/1/2012", periods=24 * 7, freq="h")
 
-    energysystem = EnergySystem(timeindex=date_time_index)
-    network.Node.registry = energysystem
+    es = EnergySystem(timeindex=date_time_index, infer_last_interval=True)
 
     # Read data file
     full_filename = os.path.join(
@@ -45,23 +43,30 @@ def test_connect_invest():
     # create electricity bus
     bel1 = Bus(label="electricity1")
     bel2 = Bus(label="electricity2")
+    es.add(bel1, bel2)
 
     # create excess component for the electricity bus to allow overproduction
-    components.Sink(label="excess_bel", inputs={bel2: Flow()})
-    components.Source(
-        label="shortage", outputs={bel2: Flow(variable_costs=50000)}
+    es.add(components.Sink(label="excess_bel", inputs={bel2: Flow()}))
+    es.add(
+        components.Source(
+            label="shortage", outputs={bel2: Flow(variable_costs=50000)}
+        )
     )
 
     # create fixed source object representing wind power plants
-    components.Source(
-        label="wind",
-        outputs={bel1: Flow(fix=data["wind"], nominal_value=1000000)},
+    es.add(
+        components.Source(
+            label="wind",
+            outputs={bel1: Flow(fix=data["wind"], nominal_capacity=1000000)},
+        )
     )
 
     # create simple sink object representing the electrical demand
-    components.Sink(
-        label="demand",
-        inputs={bel1: Flow(fix=data["demand_el"], nominal_value=1)},
+    es.add(
+        components.Sink(
+            label="demand",
+            inputs={bel1: Flow(fix=data["demand_el"], nominal_capacity=1)},
+        )
     )
 
     storage = components.GenericStorage(
@@ -74,53 +79,66 @@ def test_connect_invest():
         invest_relation_output_capacity=1 / 6,
         inflow_conversion_factor=1,
         outflow_conversion_factor=0.8,
-        investment=Investment(ep_costs=0.2),
+        nominal_capacity=Investment(ep_costs=0.2),
     )
+    es.add(storage)
 
-    line12 = components.Transformer(
+    line12 = components.Converter(
         label="line12",
         inputs={bel1: Flow()},
-        outputs={bel2: Flow(investment=Investment(ep_costs=20))},
+        outputs={bel2: Flow(nominal_capacity=Investment(ep_costs=20))},
     )
+    es.add(line12)
 
-    line21 = components.Transformer(
+    line21 = components.Converter(
         label="line21",
         inputs={bel2: Flow()},
-        outputs={bel1: Flow(investment=Investment(ep_costs=20))},
+        outputs={bel1: Flow(nominal_capacity=Investment(ep_costs=20))},
     )
+    es.add(line21)
 
-    om = Model(energysystem)
+    om = Model(es)
 
     constraints.equate_variables(
         om,
-        om.InvestmentFlowBlock.invest[line12, bel2],
-        om.InvestmentFlowBlock.invest[line21, bel1],
+        om.InvestmentFlowBlock.invest[line12, bel2, 0],
+        om.InvestmentFlowBlock.invest[line21, bel1, 0],
         2,
     )
     constraints.equate_variables(
         om,
-        om.InvestmentFlowBlock.invest[line12, bel2],
-        om.GenericInvestmentStorageBlock.invest[storage],
+        om.InvestmentFlowBlock.invest[line12, bel2, 0],
+        om.GenericInvestmentStorageBlock.invest[storage, 0],
     )
 
     # if tee_switch is true solver messages will be displayed
     logging.info("Solve the optimization problem")
-    om.solve(solver="cbc")
+    om.solve(solver="cbc", tee=True)
 
     # check if the new result object is working for custom components
     results = processing.results(om)
 
     my_results = dict()
-    my_results["line12"] = float(views.node(results, "line12")["scalars"])
-    my_results["line21"] = float(views.node(results, "line21")["scalars"])
+    my_results["line12"] = (
+        views.node(results, "line12")["scalars"]
+        .loc[[(("line12", "electricity2"), "invest")]]
+        .iloc[0]
+    )
+
+    my_results["line21"] = (
+        views.node(results, "line21")["scalars"]
+        .loc[[(("line21", "electricity1"), "invest")]]
+        .iloc[0]
+    )
+
     stor_res = views.node(results, "storage")["scalars"]
     my_results["storage_in"] = stor_res[
         [(("electricity1", "storage"), "invest")]
-    ]
-    my_results["storage"] = stor_res[[(("storage", "None"), "invest")]]
+    ].iloc[0]
+    my_results["storage"] = stor_res[[(("storage", "None"), "invest")]].iloc[0]
     my_results["storage_out"] = stor_res[
         [(("storage", "electricity1"), "invest")]
-    ]
+    ].iloc[0]
 
     connect_invest_dict = {
         "line12": 814705,
@@ -131,4 +149,6 @@ def test_connect_invest():
     }
 
     for key in connect_invest_dict.keys():
-        eq_(int(round(my_results[key])), int(round(connect_invest_dict[key])))
+        assert my_results[key] == pytest.approx(
+            connect_invest_dict[key], abs=0.5
+        )
