@@ -15,7 +15,6 @@ SPDX-License-Identifier: MIT
 import os
 
 import pandas as pd
-from oemof.network.network import Node
 from oemof.tools import economics
 
 from oemof.solph import EnergySystem
@@ -31,8 +30,6 @@ from oemof.solph.flows import Flow
 
 def test_dispatch_fix_example(solver="cbc", periods=10):
     """Invest in a flow with a `fix` sequence containing values > 1."""
-    Node.registry = None
-
     filename = os.path.join(os.path.dirname(__file__), "input_data.csv")
     data = pd.read_csv(filename, sep=",")
 
@@ -53,19 +50,23 @@ def test_dispatch_fix_example(solver="cbc", periods=10):
     pv = Source(
         label="pv",
         outputs={
-            bel: Flow(fix=data["pv"], investment=Investment(ep_costs=ep_pv))
+            bel: Flow(
+                fix=data["pv"], nominal_capacity=Investment(ep_costs=ep_pv)
+            )
         },
     )
 
     # demands (electricity/heat)
     demand_el = Sink(
         label="demand_elec",
-        inputs={bel: Flow(nominal_value=85, fix=data["demand_el"])},
+        inputs={bel: Flow(nominal_capacity=85, fix=data["demand_el"])},
     )
 
-    datetimeindex = pd.date_range("1/1/2012", periods=periods, freq="H")
+    datetimeindex = pd.date_range("1/1/2012", periods=periods, freq="h")
 
-    energysystem = EnergySystem(timeindex=datetimeindex)
+    energysystem = EnergySystem(
+        timeindex=datetimeindex, infer_last_interval=True
+    )
 
     energysystem.add(bel, excess_el, pv, demand_el)
 
@@ -80,7 +81,7 @@ def test_dispatch_fix_example(solver="cbc", periods=10):
     # ################################ results ################################
 
     # generic result object
-    results = processing.results(om=optimization_model)
+    results = processing.results(model=optimization_model)
 
     # subset of results that includes all flows into and from electrical bus
     # sequences are stored within a pandas.DataFrames and scalars e.g.
@@ -94,13 +95,3 @@ def test_dispatch_fix_example(solver="cbc", periods=10):
     comp_results["pv_capacity"] = results[(pv, bel)]["scalars"].invest
 
     assert comp_results[(("pv", "b_el"), "flow")] > 0
-
-    # test_results = {
-    #     (('pv', 'b_el'), 'flow'): 2150.5,
-    #     (('b_el', 'demand_elec'), 'flow'): 436.05,
-    #     (('b_el', 'excess_el'), 'flow'): 1714.45,
-    #     'pv_capacity': 467.5,
-    # }
-
-    # for key in test_results.keys():
-    #     eq_(int(round(comp_results[key])), int(round(test_results[key])))
