@@ -15,7 +15,7 @@ SPDX-License-Identifier: MIT
 import logging
 
 import pandas as pd
-from nose.tools import eq_
+import pytest
 
 from oemof.solph import EnergySystem
 from oemof.solph import Investment
@@ -29,12 +29,13 @@ from oemof.solph.flows import Flow
 from oemof.solph.flows import experimental as exp_flow
 
 
+@pytest.mark.skip(reason="Constraints of Electrical Line do not build.")
 def test_lopf(solver="cbc"):
     logging.info("Initialize the energy system")
 
     # create time index for 192 hours in May.
-    date_time_index = pd.date_range("5/5/2012", periods=1, freq="H")
-    es = EnergySystem(timeindex=date_time_index)
+    date_time_index = pd.date_range("5/5/2012", periods=1, freq="h")
+    es = EnergySystem(timeindex=date_time_index, infer_last_interval=True)
 
     ##########################################################################
     # Create oemof.solph objects
@@ -50,57 +51,51 @@ def test_lopf(solver="cbc"):
 
     es.add(b_el0, b_el1, b_el2)
 
-    es.add(
-        exp_flow.ElectricalLine(
-            input=b_el0,
-            output=b_el1,
-            reactance=0.0001,
-            investment=Investment(ep_costs=10),
-            min=-1,
-            max=1,
-        )
+    b_el1.inputs[b_el0] = exp_flow.ElectricalLine(
+        input=b_el0,
+        output=b_el1,
+        reactance=0.0001,
+        nominal_capacity=Investment(ep_costs=10),
+        min=-1,
+        max=1,
     )
 
-    es.add(
-        exp_flow.ElectricalLine(
-            input=b_el1,
-            output=b_el2,
-            reactance=0.0001,
-            nominal_value=60,
-            min=-1,
-            max=1,
-        )
+    b_el2.inputs[b_el1] = exp_flow.ElectricalLine(
+        input=b_el1,
+        output=b_el2,
+        reactance=0.0001,
+        nominal_capacity=60,
+        min=-1,
+        max=1,
     )
 
-    es.add(
-        exp_flow.ElectricalLine(
-            input=b_el2,
-            output=b_el0,
-            reactance=0.0001,
-            nominal_value=60,
-            min=-1,
-            max=1,
-        )
+    b_el0.inputs[b_el2] = exp_flow.ElectricalLine(
+        input=b_el2,
+        output=b_el0,
+        reactance=0.0001,
+        nominal_capacity=60,
+        min=-1,
+        max=1,
     )
 
     es.add(
         Source(
             label="gen_0",
-            outputs={b_el0: Flow(nominal_value=100, variable_costs=50)},
+            outputs={b_el0: Flow(nominal_capacity=100, variable_costs=50)},
         )
     )
 
     es.add(
         Source(
             label="gen_1",
-            outputs={b_el1: Flow(nominal_value=100, variable_costs=25)},
+            outputs={b_el1: Flow(nominal_capacity=100, variable_costs=25)},
         )
     )
 
     es.add(
         Sink(
             label="load",
-            inputs={b_el2: Flow(nominal_value=100, fix=1)},
+            inputs={b_el2: Flow(nominal_capacity=100, fix=1)},
         )
     )
 
@@ -114,7 +109,7 @@ def test_lopf(solver="cbc"):
     logging.info("Running lopf on 3-Node exmaple system")
     om.solve(solver=solver)
 
-    results = processing.results(om)
+    results = processing.results(om, remove_last_time_point=True)
 
     generators = views.node_output_by_type(results, Source)
 
@@ -125,24 +120,32 @@ def test_lopf(solver="cbc"):
 
     for key in generators_test_results.keys():
         logging.debug("Test genertor production of {0}".format(key))
-        eq_(
-            int(round(generators[key])),
-            int(round(generators_test_results[key])),
+        assert generators[key].iloc[0] == pytest.approx(
+            generators_test_results[key]
         )
 
-    eq_(
-        results[es.groups["b_2"], es.groups["b_0"]]["sequences"]["flow"][0],
-        -40,
+    assert (
+        results[es.groups["b_2"], es.groups["b_0"]]["sequences"]["flow"].iloc[
+            0
+        ]
+        == -40
     )
 
-    eq_(
-        results[es.groups["b_1"], es.groups["b_2"]]["sequences"]["flow"][0], 60
+    assert (
+        results[es.groups["b_1"], es.groups["b_2"]]["sequences"]["flow"].iloc[
+            0
+        ]
+        == 60
     )
 
-    eq_(
-        results[es.groups["b_0"], es.groups["b_1"]]["sequences"]["flow"][0],
-        -20,
+    assert (
+        results[es.groups["b_0"], es.groups["b_1"]]["sequences"]["flow"].iloc[
+            0
+        ]
+        == -20
     )
 
     # objective function
-    eq_(round(processing.meta_results(om)["objective"]), 3200)
+    assert processing.meta_results(om)["objective"] == pytest.approx(
+        3200, abs=0.5
+    )
