@@ -16,8 +16,7 @@ SPDX-License-Identifier: MIT
 import os
 
 import pandas as pd
-from nose.tools import eq_
-from oemof.network.network import Node
+import pytest
 from oemof.tools import economics
 
 from oemof.solph import EnergySystem
@@ -26,15 +25,14 @@ from oemof.solph import Model
 from oemof.solph import processing
 from oemof.solph import views
 from oemof.solph.buses import Bus
+from oemof.solph.components import Converter
 from oemof.solph.components import Sink
 from oemof.solph.components import Source
-from oemof.solph.components import Transformer
 from oemof.solph.flows import Flow
 
 
 def test_dispatch_example(solver="cbc", periods=24 * 5):
     """Create an energy system and optimize the dispatch at least costs."""
-    Node.registry = None
 
     filename = os.path.join(os.path.dirname(__file__), "input_data.csv")
     data = pd.read_csv(filename, sep=",")
@@ -63,7 +61,7 @@ def test_dispatch_example(solver="cbc", periods=24 * 5):
         outputs={
             bel: Flow(
                 fix=data["wind"],
-                investment=Investment(ep_costs=ep_wind, existing=100),
+                nominal_capacity=Investment(ep_costs=ep_wind, existing=100),
             )
         },
     )
@@ -74,7 +72,7 @@ def test_dispatch_example(solver="cbc", periods=24 * 5):
         outputs={
             bel: Flow(
                 fix=data["pv"],
-                investment=Investment(ep_costs=ep_pv, existing=80),
+                nominal_capacity=Investment(ep_costs=ep_pv, existing=80),
             )
         },
     )
@@ -82,50 +80,50 @@ def test_dispatch_example(solver="cbc", periods=24 * 5):
     # demands (electricity/heat)
     demand_el = Sink(
         label="demand_elec",
-        inputs={bel: Flow(nominal_value=85, fix=data["demand_el"])},
+        inputs={bel: Flow(nominal_capacity=85, fix=data["demand_el"])},
     )
 
     demand_th = Sink(
         label="demand_therm",
-        inputs={bth: Flow(nominal_value=40, fix=data["demand_th"])},
+        inputs={bth: Flow(nominal_capacity=40, fix=data["demand_th"])},
     )
 
     # power plants
-    pp_coal = Transformer(
+    pp_coal = Converter(
         label="pp_coal",
         inputs={bcoal: Flow()},
-        outputs={bel: Flow(nominal_value=20.2, variable_costs=25)},
+        outputs={bel: Flow(nominal_capacity=20.2, variable_costs=25)},
         conversion_factors={bel: 0.39},
     )
 
-    pp_lig = Transformer(
+    pp_lig = Converter(
         label="pp_lig",
         inputs={blig: Flow()},
-        outputs={bel: Flow(nominal_value=11.8, variable_costs=19)},
+        outputs={bel: Flow(nominal_capacity=11.8, variable_costs=19)},
         conversion_factors={bel: 0.41},
     )
 
-    pp_gas = Transformer(
+    pp_gas = Converter(
         label="pp_gas",
         inputs={bgas: Flow()},
-        outputs={bel: Flow(nominal_value=41, variable_costs=40)},
+        outputs={bel: Flow(nominal_capacity=41, variable_costs=40)},
         conversion_factors={bel: 0.50},
     )
 
-    pp_oil = Transformer(
+    pp_oil = Converter(
         label="pp_oil",
         inputs={boil: Flow()},
-        outputs={bel: Flow(nominal_value=5, variable_costs=50)},
+        outputs={bel: Flow(nominal_capacity=5, variable_costs=50)},
         conversion_factors={bel: 0.28},
     )
 
     # combined heat and power plant (chp)
-    pp_chp = Transformer(
+    pp_chp = Converter(
         label="pp_chp",
         inputs={bgas: Flow()},
         outputs={
-            bel: Flow(nominal_value=30, variable_costs=42),
-            bth: Flow(nominal_value=40),
+            bel: Flow(nominal_capacity=30, variable_costs=42),
+            bth: Flow(nominal_capacity=40),
         },
         conversion_factors={bel: 0.3, bth: 0.4},
     )
@@ -136,15 +134,17 @@ def test_dispatch_example(solver="cbc", periods=24 * 5):
     heat_source = Source(label="heat_source", outputs={b_heat_source: Flow()})
 
     cop = 3
-    heat_pump = Transformer(
+    heat_pump = Converter(
         label="el_heat_pump",
         inputs={bel: Flow(), b_heat_source: Flow()},
-        outputs={bth: Flow(nominal_value=10)},
+        outputs={bth: Flow(nominal_capacity=10)},
         conversion_factors={bel: 1 / 3, b_heat_source: (cop - 1) / cop},
     )
 
-    datetimeindex = pd.date_range("1/1/2012", periods=periods, freq="H")
-    energysystem = EnergySystem(timeindex=datetimeindex)
+    datetimeindex = pd.date_range("1/1/2012", periods=periods, freq="h")
+    energysystem = EnergySystem(
+        timeindex=datetimeindex, infer_last_interval=True
+    )
     energysystem.add(
         bcoal,
         bgas,
@@ -181,7 +181,7 @@ def test_dispatch_example(solver="cbc", periods=24 * 5):
     # ################################ results ################################
 
     # generic result object
-    results = processing.results(om=optimization_model)
+    results = processing.results(model=optimization_model)
 
     # subset of results that includes all flows into and from electrical bus
     # sequences are stored within a pandas.DataFrames and scalars e.g.
@@ -211,4 +211,4 @@ def test_dispatch_example(solver="cbc", periods=24 * 5):
     }
 
     for key in test_results.keys():
-        eq_(int(round(comp_results[key])), int(round(test_results[key])))
+        assert comp_results[key] == pytest.approx(test_results[key], abs=0.5)
